refactor(servlet): separate handle logic from HandlerMapper
